refactor: simplify docker setup
